+++ compiler/rustc_ast_lowering/src/lib.rs
@@ -88,6 +88,9 @@ mod item;
 mod lifetime_collector;
 mod pat;
 mod path;
+mod query;
+
+pub use query::provide;
 
 rustc_fluent_macro::fluent_messages! { "../messages.ftl" }
 
@@ -361,7 +364,7 @@ enum FnDeclKind {
     Impl,
 }
 
-pub fn index_ast<'tcx>(tcx: TyCtxt<'tcx>, (): ()) -> &'tcx IndexVec<LocalDefId, Steal<AstOwner>> {
+fn index_ast<'tcx>(tcx: TyCtxt<'tcx>, (): ()) -> &'tcx IndexVec<LocalDefId, Steal<AstOwner>> {
     // Queries that borrow `resolver_for_lowering`.
     tcx.ensure_with_value().output_filenames(());
     tcx.ensure_with_value().early_lint_checks(());
@@ -443,37 +446,7 @@ pub fn index_ast<'tcx>(tcx: TyCtxt<'tcx>, (): ()) -> &'tcx IndexVec<LocalDefId, 
     }
 }
 
-pub fn hir_crate<'tcx>(tcx: TyCtxt<'tcx>, (): ()) -> rustc_hir::Crate<'tcx> {
-    let mut owners: IndexVec<LocalDefId, _> = IndexVec::new();
-    while owners.next_index().index() < tcx.definitions_untracked().def_index_count() {
-        let next = owners.next_index();
-        owners.push(tcx.lower_to_hir(next));
-    }
-
-    let mut hir_body_nodes: Vec<_> = owners
-        .iter_enumerated()
-        .filter_map(|(def_id, info)| {
-            let info = info.as_owner()?;
-            let def_path_hash = tcx.hir().def_path_hash(def_id);
-            Some((def_path_hash, info))
-        })
-        .collect();
-    hir_body_nodes.sort_unstable_by_key(|bn| bn.0);
-
-    let opt_hir_hash = if tcx.needs_crate_hash() {
-        Some(tcx.with_stable_hashing_context(|mut hcx| {
-            let mut stable_hasher = StableHasher::new();
-            hir_body_nodes.hash_stable(&mut hcx, &mut stable_hasher);
-            stable_hasher.finish()
-        }))
-    } else {
-        None
-    };
-
-    rustc_hir::Crate { owners, opt_hir_hash }
-}
-
-pub fn lower_to_hir(tcx: TyCtxt<'_>, def_id: LocalDefId) -> hir::MaybeOwner<'_> {
+fn lower_to_hir(tcx: TyCtxt<'_>, def_id: LocalDefId) -> hir::MaybeOwner<'_> {
     let (resolver, _) = tcx.resolver_for_lowering();
     let ast_index = tcx.index_ast(());
     let node = ast_index.get(def_id).map(Steal::steal);
@@ -511,9 +484,37 @@ pub fn lower_to_hir(tcx: TyCtxt<'_>, def_id: LocalDefId) -> hir::MaybeOwner<'_> 
     })
 }
 
+fn hir_crate<'tcx>(tcx: TyCtxt<'tcx>, (): ()) -> rustc_hir::Crate<'tcx> {
+    let mut owners: IndexVec<LocalDefId, _> = IndexVec::new();
+    while owners.next_index().index() < tcx.definitions_untracked().def_index_count() {
+        let next = owners.next_index();
+        owners.push(tcx.lower_to_hir(next));
+    }
+
+    let mut hir_body_nodes: Vec<_> = owners
+        .iter_enumerated()
+        .filter_map(|(def_id, info)| {
+            let info = info.as_owner()?;
+            let def_path_hash = tcx.hir().def_path_hash(def_id);
+            Some((def_path_hash, info))
+        })
+        .collect();
+    hir_body_nodes.sort_unstable_by_key(|bn| bn.0);
+
+    let opt_hir_hash = if tcx.needs_crate_hash() {
+        Some(tcx.with_stable_hashing_context(|mut hcx| {
+            let mut stable_hasher = StableHasher::new();
+            hir_body_nodes.hash_stable(&mut hcx, &mut stable_hasher);
+            stable_hasher.finish()
+        }))
+    } else {
+        None
+    };
+
+    rustc_hir::Crate { owners, opt_hir_hash }
+}
+
 #[derive(Copy, Clone, PartialEq, Debug)]
-enum ParamMode {
-    /// Any path in a type context.
     Explicit,
     /// Path in a type definition, where the anonymous lifetime `'_` is not allowed.
     ExplicitNamed,
@@ -544,7 +545,6 @@ impl<'hir> LoweringContext<'hir> {
             self.tcx.hir().def_key(self.local_def_id(node_id)),
         );
 
-        let def_id = self.tcx.at(span).create_def(parent, name, def_kind).def_id();
 
         debug!("create_def: def_id_to_node_id[{:?}] <-> {:?}", def_id, node_id);
         self.node_id_to_def_id.insert(node_id, def_id);
